Replace class with a factory function
